+++ genomepy/__about__.py
@@ -1,3 +1,3 @@
 """Metadata"""
-__version__ = "0.7.1"
+__version__ = "0.7.2"
 __author__ = "Simon van Heeringen"